Move the images on qa.html to a higher location
On Firefox, the menu is too long that the Copy as Markdown item exceeds the window boundary, preventing PyAutoGUI from locating the text.

Moving the images to a higher so that the menu item is within the window's boundary.

That is, until Firefox add more items to the context menu...