Remove URLs from link titles (when present)
In particular, I also a "URL in title" extension which includes the URL in the title of every page.  This causes every markdown link to include the URL in the title text, which is not so useful.

I think this is probably useful beyond that motivating case - generally not desired to have the URL in the link title.