Attempt to build semi-e2e test with Playwright
Playwright is faster and easier to work with than Selenium.

However, it is necessary to overcome some limitations with chrome extensions.

In addition, Playwright does not work with Firefox add-ons.

The benefit of doing this is to run tests faster and without relying on OS-specific automation & configuration.

Eventually, non-critical/non-tricky functional e2e tests will be moved from Python-based tests to Playwright.

This patch was mostly done by guiding Claude Code.